# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Lotmark",
    "version": "1",
    "description": "ISO 17034 reference material producer platform.\n\nErrors are RFC 9457 problem details. The `code` field is the machine-readable one; `title` is that code humanised for display and must not be branched on. Where a refusal was recorded, `auditSeq` points at the ledger entry.\n\nThis document is generated from a registry that a test compares against the routes the server actually registered, in both directions. It cannot describe an endpoint that does not exist, and an endpoint cannot exist without appearing here."
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Local development"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "lotmark_session",
        "description": "An opaque server-side session. Cookies rather than a bearer token because this domain needs IMMEDIATE revocation — a locked account, an idle timeout, a withdrawn competence — and a stateless token cannot be withdrawn before it expires."
      }
    },
...
  },
  "paths": {
...
    "/api/v1/admin/config/draft/{id}/workflows": {
      "get": {
        "operationId": "getAdminConfigDraftByIdWorkflows",
        "summary": "The machines a draft would install",
        "description": "Resolved by the runtime resolver against the draft’s entries, and it reports the transitions the resolver DROPPED — which a diagram drawn from the raw JSON would not. Requires `user:manage`.",
        "tags": [
          "Configuration"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The id."
          }
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "The request succeeded."
          },
          "400": {
            "description": "The request was malformed. RFC 9457 problem details."
          },
          "401": {
            "description": "Not authenticated, the second factor is outstanding, or a signing step-up is required. The `code` distinguishes them."
          },
          "403": {
            "description": "Refused. `code` names the reason — permission_denied, segregation_of_duties, competence_missing or competence_expired — and `auditSeq` points at the ledger entry that recorded the refusal."
          }
        }
      }
    },
    "/api/v1/admin/config/draft/{id}/form/{entity}": {
      "get": {
        "operationId": "getAdminConfigDraftByIdFormByEntity",
        "summary": "Preview a form as this draft would render it",
        "description": "Resolved by the runtime resolver against the draft’s own entries, so the designer’s preview is the thing itself rather than a second implementation of the same rules. Requires `user:manage`.",
        "tags": [
          "Configuration"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The id."
          },
          {
            "name": "entity",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The entity."
          }
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "The request succeeded."
          },
          "400": {
            "description": "The request was malformed. RFC 9457 problem details."
          },
          "401": {
            "description": "Not authenticated, the second factor is outstanding, or a signing step-up is required. The `code` distinguishes them."
          },
          "403": {
            "description": "Refused. `code` names the reason — permission_denied, segregation_of_duties, competence_missing or competence_expired — and `auditSeq` points at the ledger entry that recorded the refusal."
          }
        }
      }
    },
...
    "/api/v1/admin/config": {
      "get": {
        "operationId": "getAdminConfig",
        "summary": "Configuration versions",
        "description": "Requires `user:manage`.",
        "tags": [
          "Administration"
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "The request succeeded."
          },
          "400": {
            "description": "The request was malformed. RFC 9457 problem details."
          },
          "401": {
            "description": "Not authenticated, the second factor is outstanding, or a signing step-up is required. The `code` distinguishes them."
          },
          "403": {
            "description": "Refused. `code` names the reason — permission_denied, segregation_of_duties, competence_missing or competence_expired — and `auditSeq` points at the ledger entry that recorded the refusal."
          }
        }
      }
    },
    "/api/v1/admin/config/{id}": {
      "get": {
        "operationId": "getAdminConfigById",
        "summary": "One version and its entries",
        "description": "Requires `user:manage`.",
        "tags": [
          "Administration"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The id."
          }
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "The request succeeded."
          },
          "400": {
            "description": "The request was malformed. RFC 9457 problem details."
          },
          "401": {
            "description": "Not authenticated, the second factor is outstanding, or a signing step-up is required. The `code` distinguishes them."
          },
          "403": {
            "description": "Refused. `code` names the reason — permission_denied, segregation_of_duties, competence_missing or competence_expired — and `auditSeq` points at the ledger entry that recorded the refusal."
          }
        }
      }
    },
    "/api/v1/admin/config/draft": {
      "post": {
        "operationId": "postAdminConfigDraft",
        "summary": "Open a configuration draft",
        "description": "One draft per tenant. Two would be a fork. Requires `user:manage`.",
        "tags": [
          "Administration"
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "The request succeeded."
          },
          "400": {
            "description": "The request was malformed. RFC 9457 problem details."
          },
          "401": {
            "description": "Not authenticated, the second factor is outstanding, or a signing step-up is required. The `code` distinguishes them."
          },
          "403": {
            "description": "Refused. `code` names the reason — permission_denied, segregation_of_duties, competence_missing or competence_expired — and `auditSeq` points at the ledger entry that recorded the refusal."
          }
        }
      }
    },
    "/api/v1/admin/config/draft/{id}/entry": {
      "put": {
        "operationId": "putAdminConfigDraftByIdEntry",
        "summary": "Write a configuration entry",
        "description": "Requires `user:manage`.",
        "tags": [
          "Administration"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The id."
          }
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "The request succeeded."
          },
          "400": {
            "description": "The request was malformed. RFC 9457 problem details."
          },
          "401": {
            "description": "Not authenticated, the second factor is outstanding, or a signing step-up is required. The `code` distinguishes them."
          },
          "403": {
            "description": "Refused. `code` names the reason — permission_denied, segregation_of_duties, competence_missing or competence_expired — and `auditSeq` points at the ledger entry that recorded the refusal."
          }
        }
      }
    },
    "/api/v1/admin/config/draft/{id}/entry/{kind}/{key}": {
      "delete": {
        "operationId": "deleteAdminConfigDraftByIdEntryByKindByKey",
        "summary": "Remove a configuration entry",
        "description": "Requires `user:manage`.",
        "tags": [
          "Administration"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The id."
          },
          {
            "name": "kind",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The kind."
          },
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The key."
          }
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "The request succeeded."
          },
          "400": {
            "description": "The request was malformed. RFC 9457 problem details."
          },
          "401": {
            "description": "Not authenticated, the second factor is outstanding, or a signing step-up is required. The `code` distinguishes them."
          },
          "403": {
            "description": "Refused. `code` names the reason — permission_denied, segregation_of_duties, competence_missing or competence_expired — and `auditSeq` points at the ledger entry that recorded the refusal."
          }
        }
      }
    },
    "/api/v1/admin/config/draft/{id}/review": {
      "get": {
        "operationId": "getAdminConfigDraftByIdReview",
        "summary": "The diff, and everything that would refuse publication",
        "description": "Requires `user:manage`.",
        "tags": [
          "Administration"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The id."
          }
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "The request succeeded."
          },
          "400": {
            "description": "The request was malformed. RFC 9457 problem details."
          },
          "401": {
            "description": "Not authenticated, the second factor is outstanding, or a signing step-up is required. The `code` distinguishes them."
          },
          "403": {
            "description": "Refused. `code` names the reason — permission_denied, segregation_of_duties, competence_missing or competence_expired — and `auditSeq` points at the ledger entry that recorded the refusal."
          }
        }
      }
    },
    "/api/v1/admin/config/draft/{id}/publish": {
      "post": {
        "operationId": "postAdminConfigDraftByIdPublish",
        "summary": "Publish a draft",
        "description": "A security or behaviour change must be signed. Publishability is checked BEFORE the signature is collected. Requires `user:manage`.",
        "tags": [
          "Administration"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The id."
          }
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "The request succeeded."
          },
          "400": {
            "description": "The request was malformed. RFC 9457 problem details."
          },
          "401": {
            "description": "Not authenticated, the second factor is outstanding, or a signing step-up is required. The `code` distinguishes them."
          },
          "403": {
            "description": "Refused. `code` names the reason — permission_denied, segregation_of_duties, competence_missing or competence_expired — and `auditSeq` points at the ledger entry that recorded the refusal."
          }
        }
      }
    },
    "/api/v1/admin/config/draft/{id}": {
      "delete": {
        "operationId": "deleteAdminConfigDraftById",
        "summary": "Discard a draft",
        "description": "Requires `user:manage`.",
        "tags": [
          "Administration"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The id."
          }
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "The request succeeded."
          },
          "400": {
            "description": "The request was malformed. RFC 9457 problem details."
          },
          "401": {
            "description": "Not authenticated, the second factor is outstanding, or a signing step-up is required. The `code` distinguishes them."
          },
          "403": {
            "description": "Refused. `code` names the reason — permission_denied, segregation_of_duties, competence_missing or competence_expired — and `auditSeq` points at the ledger entry that recorded the refusal."
          }
        }
      }
    },
...
  }
}